requirements.txt
# Function dependencies, for example:
# package>=version
functions_framework
google-cloud-bigquery-datatransfer